% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_possibilities.R
\name{all_possibilities}
\alias{all_possibilities}
\title{all_possibilities}
\usage{
all_possibilities(game)
}
\arguments{
\item{game}{chess game object (i.e., a list with elements board, turn, history, and fen_history
as created by newgame function)}
}
\value{
all chess possibilities
}
\description{
finds all legal moves
}
\examples{
newgame() |>
all_possibilities()

}
