% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess2plyrs_Package.R
\docType{package}
\name{chess2plyrs-package}
\alias{chess2plyrs}
\alias{chess2plyrs-package}
\title{chess2plyrs: Chess Game and Tools in R}
\description{
Chess package written in base R. The program allows the user to create a game, add moves, check for legal moves and game result, plot the board, take back, read and write FEN. A basic chess engine are implemented.
}
\author{
\strong{Maintainer}: Luigi Annicchiarico \email{luigi.annic@gmail.com}

}
\keyword{internal}
