% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_move.R
\name{chess_move}
\alias{chess_move}
\title{chess_move}
\usage{
chess_move(game, piece, initialposition = "", finalposition = "")
}
\arguments{
\item{game}{chess game object (i.e., a list with elements board, turn, history, and fen_history
as created by newgame function)}

\item{piece}{letter indicating the piece to be moved (p, N, B, R, Q, K)}

\item{initialposition}{initial square of the piece}

\item{finalposition}{destination square}
}
\value{
makes move
}
\description{
Takes in input a move, evaluates whether it is legal,
            and if it is, then the game is updated
}
\examples{
newgame() |>
chess_move("N", "g1", "f3")

}
