% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chesstools.R
\docType{data}
\name{chesstools}
\alias{chesstools}
\title{chesstools}
\format{
An object of class \code{list}
}
\usage{
chesstools
}
\description{
All tools used for letting the chess program work.
}
\keyword{list}
