% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_score1.R
\name{engine1}
\alias{engine1}
\title{engine1}
\usage{
engine1(game, depth)
}
\arguments{
\item{game}{chess game object (i.e., a list with elements board, turn, history, and fen_history
as created by newgame function)}

\item{depth}{depth of the minimax. depth of 1 and 2 are fairly rapid.}
}
\value{
game with new move done
}
\description{
engine which chooses minimax between legal moves
}
