% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_result.R
\name{game_result}
\alias{game_result}
\title{game_result}
\usage{
game_result(currentboard = game$board, turn = game$turn)
}
\arguments{
\item{currentboard}{game$currentboard}

\item{turn}{game4turn}
}
\value{
a message
}
\description{
# game ending check
}
