% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_score1.R
\name{get_minimax_move}
\alias{get_minimax_move}
\title{get_minimax_move}
\usage{
get_minimax_move(game, depth)
}
\arguments{
\item{game}{chess game object (i.e., a list with elements board, turn, history, and fen_history
as created by newgame function)}

\item{depth}{algorithm depth}
}
\value{
minimax engine
}
\description{
minimax engine
}
