% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legalmoves.R
\name{legalmoves}
\alias{legalmoves}
\title{legalmoves}
\usage{
legalmoves(turn = game$turn)
}
\arguments{
\item{turn}{turn}
}
\value{
character vector
}
\description{
lists legal moves
}
\examples{
game <- newgame()
legalmoves()

}
