% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_move4.R
\name{make_move4}
\alias{make_move4}
\title{make_move4}
\usage{
make_move4(game, piece, initialposition = "", finalposition = "")
}
\arguments{
\item{game}{chess game object (i.e., a list with elements board, turn, history, and fen_history
as created by newgame function)}

\item{piece}{piece to be moved}

\item{initialposition}{initial square of the piece}

\item{finalposition}{final square}
}
\value{
makes move
}
\description{
makes move
}
