% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_mover.R
\name{random_mover}
\alias{random_mover}
\title{random_mover}
\usage{
random_mover(game)
}
\arguments{
\item{game}{chess game object (i.e., a list with elements board, turn, history, and fen_history
as created by newgame function)}
}
\value{
game with new move done
}
\description{
engine which chooses randomly between legal moves
}
