% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readfen.R
\name{readfen}
\alias{readfen}
\title{readfen}
\usage{
readfen(fenstring)
}
\arguments{
\item{fenstring}{fen string (pieces and turn)}
}
\value{
board and turn
}
\description{
read fen (Forsyth–Edwards Notation) notation
}
\examples{
readfen("r3kb1r/pp1nqppp/4bp2/2p5/3P4/1B3N2/PPP1QPPP/R1B1K2R w")

}
