% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/takeback.R
\name{takeback}
\alias{takeback}
\title{takeback}
\usage{
takeback(game)
}
\arguments{
\item{game}{chess game object (i.e., a list with elements board, turn, history, and fen_history
as created by newgame function)}
}
\value{
game (last move being deleted)
}
\description{
takeback
}
