% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefen.R
\name{writefen}
\alias{writefen}
\title{writefen}
\usage{
writefen(currentboard = game$board, turn = game$turn)
}
\arguments{
\item{currentboard}{currentboard}

\item{turn}{game$turn}
}
\value{
fen
}
\description{
writes fen
}
