% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_possibilities.R
\name{all_possibilities}
\alias{all_possibilities}
\title{all_possibilities}
\usage{
all_possibilities(
  currentboard = game$board,
  turn = game$turn,
  history = game$history
)
}
\arguments{
\item{currentboard}{game$board}

\item{turn}{turn}

\item{history}{history}
}
\value{
all chess possibilities
}
\description{
finds all legal moves
}
\examples{
game <- newgame()
all_possibilities()

}
