% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_possibilities.R
\name{all_possibilities}
\alias{all_possibilities}
\title{all_possibilities}
\usage{
all_possibilities(currentboard = game$board)
}
\arguments{
\item{currentboard}{game$board}
}
\value{
all chess possiblities
}
\description{
finds all legal moves
}
\examples{
game <- newgame()
all_possibilities()

}
