% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess2plyrs_Package.R
\docType{package}
\name{chess2plyrs-package}
\alias{chess2plyrs}
\alias{chess2plyrs-package}
\title{chess2plyrs: Chess player in base R}
\description{
Chess program for playing against a friend, and (hopefully) soon a chess engine implemented.
}
\author{
\strong{Maintainer}: Luigi Annicchiarico \email{luigi.annic@gmail.com}

}
\keyword{internal}
