% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess2plyrs_Package.R
\docType{package}
\name{chess2plyrs-package}
\alias{chess2plyrs}
\alias{chess2plyrs-package}
\title{chess2plyrs: Chess Game Creation and Tools}
\description{
A chess program which allows the user to create a game, add moves, check for legal moves and game result, plot the board, take back, read and write FEN (Forsyth–Edwards Notation). A basic chess engine based on minimax is implemented.
}
\author{
\strong{Maintainer}: Luigi Annicchiarico \email{luigi.annic@gmail.com}

}
\keyword{internal}
