% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess2plyrs_Package.R
\docType{package}
\name{chess2plyrs-package}
\alias{chess2plyrs}
\alias{chess2plyrs-package}
\title{chess2plyrs: Chess player in base R}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
