% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess2plyrs_Package.R
\docType{package}
\name{chess2plyrs-package}
\alias{chess2plyrs}
\alias{chess2plyrs-package}
\title{chess2plyrs: Chess for 2 Players}
\description{
Chess program for playing against a friend. The program checks for legal moves and provides the possibility to plot the board, take back, read and write FEN, and list all legal moves in the current position. Some simple chess engines are implemented.
}
\author{
\strong{Maintainer}: Luigi Annicchiarico \email{luigi.annic@gmail.com}

}
\keyword{internal}
