% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chessplot.R
\name{chessplot}
\alias{chessplot}
\title{chessplot}
\usage{
chessplot(game)
}
\arguments{
\item{game}{chess game object (i.e., a list with elements board, turn, history, and fen_history
as created by newgame function)}
}
\value{
plot
}
\description{
chess plot
}
