% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_result.R
\name{game_result}
\alias{game_result}
\title{game_result}
\usage{
game_result(game)
}
\arguments{
\item{game}{chess game object (i.e., a list with elements board, turn, history, and fen_history
as created by newgame function)}
}
\value{
a message
}
\description{
# game ending check
}
