% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_move4.R
\name{make_move4}
\alias{make_move4}
\title{make_move4}
\usage{
make_move4(
  piece,
  initialposition = "",
  finalposition = "",
  currentboard = game$board,
  turn = game$turn,
  history = game$history
)
}
\arguments{
\item{piece}{piece to be moved}

\item{initialposition}{initial square of the piece}

\item{finalposition}{final square}

\item{currentboard}{game$board}

\item{turn}{game$turn}

\item{history}{history}
}
\value{
makes move
}
\description{
makes move
}
