% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moves_scoresheet.R
\name{moves_scoresheet}
\alias{moves_scoresheet}
\title{moves_scoresheet}
\usage{
moves_scoresheet(h = game$history, shortnotation = TRUE)
}
\arguments{
\item{h}{game$hstory}

\item{shortnotation}{Use short scientific notation? TRUE is he default}
}
\value{
moves scoresheet
}
\description{
Creates move scorelist, in beautiful notation
}
