% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefen.R
\name{writefen}
\alias{writefen}
\title{writefen}
\usage{
writefen(game, cb = NULL, tb = NULL, cb_tb_insteadof_game = FALSE)
}
\arguments{
\item{game}{chess game object (i.e., a list with elements board, turn, history, and fen_history
as created by newgame function)}

\item{cb}{chess board if cb_tf_insteadof_game set to TRUE}

\item{tb}{turn if cb_tf_insteadof_game set to TRUE}

\item{cb_tb_insteadof_game}{if FALSE, uses game to create fen, if TRUE it uses cb and tb}
}
\value{
fen
}
\description{
write fen
}
